igvm: Relax unsafe_code restriction
Currently unsafe_code is forbidden for igvm code. In order to add
a C compatible API we will need to accept and work with raw C
pointers to data. This requires a relaxing of the unsafe rule from
'forbid' to 'deny' allowing the C API feature, when enabled,  to
override the setting.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>