igvm: Add header_type() function to variable headers
The C API will need to convert variable headers from the rust
representation to C compatible structures. As part of this the
C API will need to determine the original IgvmVariableHeaderType
from the rust representation.

This patch adds the required functions behind the igvm-c feature
as they are not used outside of that feature.

Signed-off-by: Roy Hopkins <rhopkins@suse.de>