igvm_defs: Remove INVALID from IgvmVariableHeaderType enum
The IgvmVariableHeaderType enum includes an entry for INVALID defined as
0. However, any value other than those defined by the enumeration should
be deemed invalid.

Additionally, the INVALID entry does not conform to the convention use
by the other enumeration values (starting with IGVM_VHT_) and will cause
a name clash with INVALID entries in other enumerations when generating a
C API from this code.

Therefore, this commit removes the INVALID enumeration.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>