Generate a C API for IGVM (#3)
This is a proposal on how to implement
https://github.com/microsoft/igvm/issues/1 and provide a C API for igvm.

The approach taken is to based on using
[cbindgen](https://github.com/mozilla/cbindgen) to analyze the rust code
and produce header files usable from C. To support this, a set of C
friendly functions have been added to `igvm/src/c_api.rs` which allows,
at present, binary files to be parsed and processed.

The additional functions are hidden behind a rust feature named
`igvm-c`. When performing a normal build (without the C API) the only
change that has any effect is from the first commit
4da1043900c54dd8f5a47c46df5d63d1f402453b which by necessity relaxes the
unsafe_code restriction. Unfortunately this cannot be controlled by the
feature.

The sample in `igvm_c/sample` and the unit test code in `igvm/tests` are
probably the best places to start to see how to use the proposed API
from C.

---------

Signed-off-by: Roy Hopkins <rhopkins@suse.de>
Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>