igvm_c: Fix linker static libaries for tests and dump_igvm
The static library built by rustc depnds on a number of static
libraries that must be linked into the resulting executable.
Depending on the compiler/environment, this caused an error
when building the C unit tests and dump_igvm in some cases.

This commit adds the required static libraries to the
Makefile.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>