igvm_c: Ensure C header types are prefixed with IGVM in igvm_defs.h
The generated header file igvm_defs.h contains a number of definitions
that could potentially clash with other non-IGVM header files, for
example "u32_le". This change ensures that all types defined in the
generated header file are prefixed with IGVM/Igvm.

A new post-processing script needed to added to implement some of these
changes as cbindgen does not have the granularity to make the necessary
changes. In addition, some annototations needed to be added to the rust
code itself.

Note that this is a breaking change to the header file, but a necessary
one that is being introduced before adoption of the IGVM C API.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>