igvm: Add SEV and SEV_ES platform types
Currently, only SEV-SNP is supported as a platform type for SEV
guests. However, there is no reason why IGVM cannot be used to
configure both SEV and SEV-ES.

This commit adds two new platform types: `IgvmPlatformType::SEV` and
`IgvmPlatformType::SEV_ES` which allow IGVM files to specifically target
these isolation technologies.

Like SEV-SNP, SEV-ES allows the VP context to be specified which affects
the initial guest measurement. The structure for defining the context
is binary compatible with SEV-SNP, therefore the SEV-ES platform type
reuses IgvmDirectiveHeader::SnpVpContext to define the VMSA.

In contrast, an IGVM loader can apply an initial CPU state for SEV but
this is not included in the guest measurement. Therefore the
IgvmNativeVpContextX64 structure is used to describe the context in this
case.

These are added as an unstable feature for V2. However, the definitions
will be included in the generated C API header files by default.

Signed-off-by: Roy Hopkins <roy.hopkins@suse.com>