igvm_c: Templatize pkgconfig file and set correct crate version
Turn igvm.pc into a template file (igvm.pc.in) that is expanded
and written to its destination during the 'install' Makefile target.
This is more robust than the current approach (separate target).

Update the Version field in the pkgconfig file and set it to the
current version of the crate automatically.

Signed-off-by: Oliver Steffen <osteffen@redhat.com>