igvm_defs: add support for specific purpose memory in MemoryMapEntryType. (#64)
This PR adds a MemoryMapEntryType for specific purpose memory. This
corresponds with the UEFI memory map attribute EFI_MEMORY_SP, added in
UEFI 2.8. EFI_MEMORY_SP is used to indicate that the memory has special
properties and is to be reserved for application use, and shouldn't be
used by the firmware or OS.

---------

Co-authored-by: Patrick Payne <patpa@microsoft.com>